added support for explicit RX subprocess execution, along with interactive subprocess support for same use case

+++ src/main/kotlin/dev/specter/ingen/config/ConfigBuilder.kt
@@ -6,10 +6,12 @@ import dev.specter.ingen.config.IngenDefaults.CMD_PATH
 import dev.specter.ingen.config.IngenDefaults.CONFIG_PATH
 import dev.specter.ingen.config.IngenDefaults.MODULE_1_PATH
 import dev.specter.ingen.config.IngenDefaults.MODULE_2_PATH
+import dev.specter.ingen.config.IngenDefaults.MODULE_3_PATH
 import dev.specter.ingen.util.CommandConstants
 import dev.specter.ingen.util.Logger
 import dev.specter.ingen.util.SerializationHandler
 import java.nio.file.Files
+import kotlin.io.path.exists
 
 /**
  * Utility object for quickly parsing subprocess commands from a given JSON file
@@ -28,27 +30,41 @@ object ConfigBuilder {
                     "\tLog dir: $logDirCreated\n" +
                     "\tModule dir: $moduleDirCreated\n")
         }
+        generateDefaultFiles()
     }
 
     // TODO: abstract this type of file creation and overwriting in Auxi
     fun generateDefaultFiles(): Boolean = try {
-        Files.write(
-            CMD_PATH,
-            IngenDefaults.DEFAULT_COMMANDS.toByteArray(),
-        )
-        Files.write(
-            CONFIG_PATH,
-            IngenDefaults.DEFAULT_CONFIG.toByteArray(),
-        )
-        Files.write(
-            MODULE_1_PATH,
-            ScriptDefaults.INTERACTIVE_PY_TEST_SCRIPT.toByteArray(),
-        )
-        Files.write(
-            MODULE_2_PATH,
-            ScriptDefaults.ASYNC_EMITTER_SH_TEST_SCRIPT.toByteArray(),
-        )
-
+        if (CMD_PATH.exists().not()) {
+            Files.write(
+                CMD_PATH,
+                IngenDefaults.DEFAULT_COMMANDS.toByteArray(),
+            )
+        }
+        if (CONFIG_PATH.exists().not()) {
+            Files.write(
+                CONFIG_PATH,
+                IngenDefaults.DEFAULT_CONFIG.toByteArray(),
+            )
+        }
+        if (MODULE_1_PATH.exists().not()) {
+            Files.write(
+                MODULE_1_PATH,
+                ScriptDefaults.INTERACTIVE_PY_TEST_SCRIPT.toByteArray(),
+            )
+        }
+        if (MODULE_2_PATH.exists().not()) {
+            Files.write(
+                MODULE_2_PATH,
+                ScriptDefaults.ASYNC_EMITTER_SH_TEST_SCRIPT.toByteArray(),
+            )
+        }
+        if (MODULE_3_PATH.exists().not()) {
+            Files.write(
+                MODULE_3_PATH,
+                ScriptDefaults.ASYNC_EMITTER_PY_TEST_SCRIPT.toByteArray(),
+            )
+        }
         true
     } catch (e: Exception) {
         Logger.error(e)
